+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## staged
 
+- Added capability to infer number of phases from supported dss objects
 - Fixed bug in `calc_connected_components` where function signature was incorrect
 
 ## v0.12.0